refactor: refatorando código anterior dos karts
